Add improvements
