+++ extra/man/stackablectl.1
@@ -2,11 +2,11 @@
 .el .ds Aq '
 .TH stackablectl 1  "stackablectl 0.0.0-dev" 
 .SH NAME
-stackablectl \- Command line tool to interact with a Stackable Data Platform
+stackablectl \- Command line tool to interact with the Stackable Data Platform
 .SH SYNOPSIS
 \fBstackablectl\fR [\fB\-l\fR|\fB\-\-log\-level\fR] [\fB\-\-no\-cache\fR] [\fB\-\-offline\fR] [\fB\-d\fR|\fB\-\-demo\-file\fR] [\fB\-s\fR|\fB\-\-stack\-file\fR] [\fB\-r\fR|\fB\-\-release\-file\fR] [\fB\-\-helm\-repo\-stable\fR] [\fB\-\-helm\-repo\-test\fR] [\fB\-\-helm\-repo\-dev\fR] [\fB\-h\fR|\fB\-\-help\fR] [\fB\-V\fR|\fB\-\-version\fR] <\fIsubcommands\fR>
 .SH DESCRIPTION
-Command line tool to interact with a Stackable Data Platform
+Command line tool to interact with the Stackable Data Platform
 .SH OPTIONS
 .TP
 \fB\-l\fR, \fB\-\-log\-level\fR=\fILOG_LEVEL\fR
