+++ rust/stackablectl/README.md
@@ -10,7 +10,7 @@ The `stackablectl` binary provides extensive help and usage information. This he
 as any subcommands using the `--help` flag.
 
 ```plain
-Command line tool to interact with a Stackable Data Platform
+Command line tool to interact with the Stackable Data Platform
 
 Usage: stackablectl [OPTIONS] <COMMAND>
